requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
basicauth==0.4.1
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
configparser==5.0.0
cryptography==2.9.2
Deprecated==1.2.10
get==2019.4.13
github3.py==1.3.0
h11==0.8.1
h2==3.2.0
hpack==3.0.0
http3==0.6.7
hyperframe==5.2.0
idna==2.9
idna-ssl==1.1.0
jwcrypto==0.7
multidict==4.7.6
post==2019.4.13
public==2019.4.13
pycparser==2.20
PyGithub==1.51
PyJWT==1.7.1
python-dateutil==2.8.1
query-string==2019.4.13
request==2019.4.13
requests==2.23.0
requests-async==0.6.2
rfc3986==1.4.0
six==1.15.0
syncer==1.3.0
threadpool==1.3.2
typing-extensions==3.7.4.2
uritemplate==3.0.1
urllib3==1.25.9
wrapt==1.12.1
yarl==1.4.2
